refactor(cmd): apply code-review fixes — rename location_str, extract test helpers
- Rename _location_str → location_str (was exported but underscore-named)
- Rename raw_lines → raws in TreePipeline.run() to match bytes framing in docstring
- Add comment on stderr=None in Popen calls to clarify intent
- Extract _proc, _tree_record, _error_record, _error_record_with_source,
  _eof_error_record into _test_helpers.py; remove duplicates from three test files
