+++ src/plcc/cmd/rep_test.py
@@ -1,5 +1,4 @@
 import io
-import json
 import subprocess
 import sys
 from types import SimpleNamespace
@@ -7,42 +6,10 @@ from types import SimpleNamespace
 import pytest
 
 from .rep import RepHandler
-
-
-def _proc(stdout=b"", returncode=0):
-    p = SimpleNamespace(returncode=returncode)
-    p.communicate = lambda: (stdout, b"")
-    p.wait = lambda: None
-    p.stdin = io.BytesIO()
-    p.stdout = io.BytesIO(stdout)
-    return p
-
-
-def _tree():
-    return json.dumps({
-        "kind": "tree", "rule": "program",
-        "source": {}, "children": []
-    }).encode() + b"\n"
-
-
-def _error_record(msg="bad", stage="plcc-tokens"):
-    return json.dumps({"kind": "error", "message": msg, "stage": stage}).encode() + b"\n"
-
-
-def _error_record_with_source(msg="bad", stage="plcc-parser-table",
-                               file="-", line=2, col=5):
-    return json.dumps({
-        "kind": "error", "message": msg, "stage": stage,
-        "source": {"file": file, "line": line, "column": col},
-    }).encode() + b"\n"
-
-
-def _eof_error_record(msg="unexpected end of input"):
-    return json.dumps({
-        "kind": "error", "message": msg,
-        "found": "eof",
-        "source": {"file": "-", "line": 1, "column": 1},
-    }).encode() + b"\n"
+from ._test_helpers import (
+    _proc, _tree_record, _error_record, _error_record_with_source,
+    _eof_error_record,
+)
 
 
 def _make_interpreter(response=b'{"kind":"result","value":"42"}\n'):
@@ -72,7 +39,7 @@ def test_feed_returns_false_when_no_tree(monkeypatch, handler):
 
 def test_feed_returns_true_when_tree_produced(monkeypatch, handler):
     h, _ = handler
-    procs = iter([_proc(), _proc(stdout=_tree())])
+    procs = iter([_proc(), _proc(stdout=_tree_record())])
     monkeypatch.setattr(subprocess, "Popen", lambda *a, **kw: next(procs))
     assert h.feed(b"42\n", "-") is True
 
@@ -87,7 +54,7 @@ def test_feed_does_not_contact_interpreter_when_no_tree(monkeypatch, handler):
 
 def test_feed_sends_tree_to_interpreter(monkeypatch, handler):
     h, interp = handler
-    procs = iter([_proc(), _proc(stdout=_tree())])
+    procs = iter([_proc(), _proc(stdout=_tree_record())])
     monkeypatch.setattr(subprocess, "Popen", lambda *a, **kw: next(procs))
     h.feed(b"42\n", "-")
     written = interp.stdin.getvalue()
@@ -96,7 +63,7 @@ def test_feed_sends_tree_to_interpreter(monkeypatch, handler):
 
 def test_feed_prints_result(monkeypatch, handler, capsys):
     h, _ = handler
-    procs = iter([_proc(), _proc(stdout=_tree())])
+    procs = iter([_proc(), _proc(stdout=_tree_record())])
     monkeypatch.setattr(subprocess, "Popen", lambda *a, **kw: next(procs))
     h.feed(b"42\n", "-")
     out, _ = capsys.readouterr()
@@ -122,7 +89,7 @@ def test_feed_handles_non_dict_json_number_from_interpreter(monkeypatch, handler
     response = b'42\n{"kind":"result","value":"ok"}\n'
     h, _ = handler
     h._interpreter.stdout = io.BytesIO(response)
-    procs = iter([_proc(), _proc(stdout=_tree())])
+    procs = iter([_proc(), _proc(stdout=_tree_record())])
     monkeypatch.setattr(subprocess, "Popen", lambda *a, **kw: next(procs))
     h.feed(b"1\n", "-")
     out, _ = capsys.readouterr()
@@ -135,7 +102,7 @@ def test_feed_handles_non_dict_json_array_from_interpreter(monkeypatch, handler,
     response = b'[1,2,3]\n{"kind":"result","value":"done"}\n'
     h, _ = handler
     h._interpreter.stdout = io.BytesIO(response)
-    procs = iter([_proc(), _proc(stdout=_tree())])
+    procs = iter([_proc(), _proc(stdout=_tree_record())])
     monkeypatch.setattr(subprocess, "Popen", lambda *a, **kw: next(procs))
     h.feed(b"1\n", "-")
     out, _ = capsys.readouterr()
